Extract Cluster interface
